Extracted refactoring of NP map
